fix overflow issue

+++ corelib/dynamicemb/src/unique_op.cu
@@ -397,7 +397,8 @@ segmented_unique_kernel(const KeyType *d_keys, const int32_t *d_table_ids,
               static_cast<int32_t>(atomicAdd(&table_counters[table_id], 1));
 
           // Store unique key in partitioned layout
-          size_t output_pos = table_id * max_keys_per_table + local_unique_idx;
+          size_t output_pos =
+              static_cast<size_t>(table_id) * max_keys_per_table + local_unique_idx;
           d_unique_keys[output_pos] = key;
 
           // Pack and store (table_id, local_idx) - this signals completion
@@ -429,7 +430,8 @@ segmented_unique_kernel(const KeyType *d_keys, const int32_t *d_table_ids,
 
             // Update frequency counter for duplicate key
             if (frequency_counters) {
-              size_t output_pos = table_id * max_keys_per_table + local_idx;
+              size_t output_pos =
+                  static_cast<size_t>(table_id) * max_keys_per_table + local_idx;
               atomicAdd(&frequency_counters[output_pos], input_freq);
             }
             done = true;
@@ -453,7 +455,8 @@ segmented_unique_kernel(const KeyType *d_keys, const int32_t *d_table_ids,
 
           // Update frequency counter for duplicate key
           if (frequency_counters) {
-            size_t output_pos = table_id * max_keys_per_table + local_idx;
+            size_t output_pos =
+                static_cast<size_t>(table_id) * max_keys_per_table + local_idx;
             atomicAdd(&frequency_counters[output_pos], input_freq);
           }
           done = true;
@@ -506,7 +509,7 @@ __global__ void compact_keys_and_freq_kernel(
     int64_t local_idx = idx - table_offsets[table_id];
 
     // Compute source position in partitioned layout
-    size_t src_pos = table_id * max_keys_per_table + local_idx;
+    size_t src_pos = static_cast<size_t>(table_id) * max_keys_per_table + local_idx;
 
     // Compact keys
     output_keys[idx] = partitioned_keys[src_pos];
